+++ Platform/OpenCore/OpenCoreUefi.c
@@ -446,9 +446,11 @@ OcLoadBooterUefiSupport (
   IN OC_GLOBAL_CONFIG  *Config
   )
 {
-  OC_ABC_SETTINGS  AbcSettings;
-  UINT32           Index;
-  UINT32           NextIndex;
+  OC_ABC_SETTINGS        AbcSettings;
+  UINT32                 Index;
+  UINT32                 NextIndex;
+  OC_BOOTER_PATCH_ENTRY  *UserPatch;
+  OC_BOOTER_PATCH        *Patch;
 
   ZeroMem (&AbcSettings, sizeof (AbcSettings));
 
@@ -471,6 +473,9 @@ OcLoadBooterUefiSupport (
   AbcSettings.SignalAppleOS          = Config->Booter.Quirks.SignalAppleOS;
   AbcSettings.SyncRuntimePermissions = Config->Booter.Quirks.SyncRuntimePermissions;
 
+  //
+  // Handle MmioWhitelist patches.
+  //
   if (AbcSettings.DevirtualiseMmio && Config->Booter.MmioWhitelist.Count > 0) {
     AbcSettings.MmioWhitelist = AllocatePool (
       Config->Booter.MmioWhitelist.Count * sizeof (AbcSettings.MmioWhitelist[0])
@@ -494,12 +499,97 @@ OcLoadBooterUefiSupport (
     }
   }
 
+  //
+  // Handle customised booter patches.
+  //
+  if (Config->Booter.Patch.Count > 0) {
+    AbcSettings.BooterPatches = AllocateZeroPool (
+      Config->Booter.Patch.Count * sizeof (AbcSettings.BooterPatches[0])
+      );
+
+    if (AbcSettings.BooterPatches != NULL) {
+      NextIndex = 0;
+      for (Index = 0; Index < Config->Booter.Patch.Count; ++Index) {
+        Patch = &AbcSettings.BooterPatches[NextIndex];
+        UserPatch = Config->Booter.Patch.Values[Index];
+
+        if (!UserPatch->Enabled) {
+          continue;
+        }
+
+        //
+        // Ignore patch if:
+        // - There is nothing to replace.
+        // - Find and replace mismatch in size.
+        // - Mask and ReplaceMask mismatch in size when are available.
+        //
+        if (UserPatch->Replace.Size == 0
+          || UserPatch->Find.Size != UserPatch->Replace.Size
+          || (UserPatch->Mask.Size > 0 && UserPatch->Find.Size != UserPatch->Mask.Size)
+          || (UserPatch->ReplaceMask.Size > 0 && UserPatch->Find.Size != UserPatch->ReplaceMask.Size)) {
+          DEBUG ((DEBUG_ERROR, "OC: Booter patch %u is borked\n", Index));
+          continue;
+        }
+
+        //
+        // Also, ignore patch on mismatched architecture.
+        //
+        Patch->Arch    = OC_BLOB_GET (&UserPatch->Arch);
+        if (Patch->Arch[0] != '\0' && AsciiStrCmp (Patch->Arch, "Any") != 0) {
+#if defined(MDE_CPU_X64)
+          if (AsciiStrCmp (Patch->Arch, "x86_64") != 0) {
+            continue;
+          }
+#elif defined(MDE_CPU_IA32)
+          if (AsciiStrCmp (Patch->Arch, "i386") != 0) {
+            continue;
+          }
+#else
+#error "Unsupported architecture"
+#endif
+        }
+
+        //
+        // Here we simply receive Identifier from user,
+        // and it will be parsed by the internal patch function.
+        //
+        Patch->Identifier = OC_BLOB_GET (&UserPatch->Identifier);
+
+        Patch->Find       = OC_BLOB_GET (&UserPatch->Find);
+        Patch->Replace    = OC_BLOB_GET (&UserPatch->Replace);
+
+        Patch->Comment    = OC_BLOB_GET (&UserPatch->Comment);
+        
+        if (UserPatch->Mask.Size > 0) {
+          Patch->Mask     = OC_BLOB_GET (&UserPatch->Mask);
+        }
+
+        if (UserPatch->ReplaceMask.Size > 0) {
+          Patch->ReplaceMask = OC_BLOB_GET (&UserPatch->ReplaceMask);
+        }
+
+        Patch->Size          = UserPatch->Replace.Size;
+        Patch->Count         = UserPatch->Count;
+        Patch->Skip          = UserPatch->Skip;
+        Patch->Limit         = UserPatch->Limit;
+
+        ++NextIndex;
+      }
+      AbcSettings.BooterPatchesSize = NextIndex;
+    } else {
+      DEBUG ((
+        DEBUG_ERROR,
+        "OC: Failed to allocate %u slots for user booter patches\n",
+        (UINT32) Config->Booter.Patch.Count
+        ));
+    }
+  }
+
   AbcSettings.ExitBootServicesHandlers = mOcExitBootServicesHandlers;
   AbcSettings.ExitBootServicesHandlerContexts = mOcExitBootServicesContexts;
 
   OcAbcInitialize (&AbcSettings);
 }
-
 
 VOID
 OcReserveMemory (